+++ src/Stack/Types/Build.hs
@@ -119,8 +119,6 @@ data StackBuildException
   | InvalidFlagSpecification (Set UnusedFlags)
   | InvalidGhcOptionsSpecification [PackageName]
   | TargetParseException [Text]
-  | SolverGiveUp String
-  | SolverMissingCabalInstall
   | SomeTargetsNotBuildable [(PackageName, NamedComponent)]
   | TestSuiteExeMissing Bool String String String
   | CabalCopyFailed Bool String
@@ -288,15 +286,6 @@ instance Show StackBuildException where
         $ "The following errors occurred while parsing the build targets:"
         : map (("- " ++) . T.unpack) errs
 
-    show (SolverGiveUp msg) = concat
-        [ "\nSolver could not resolve package dependencies.\n"
-        , "You can try the following:\n"
-        , msg
-        ]
-    show SolverMissingCabalInstall = unlines
-        [ "Solver requires that cabal be on your PATH"
-        , "Try running 'stack install cabal-install'"
-        ]
     show (SomeTargetsNotBuildable xs) =
         "The following components have 'buildable: False' set in the cabal configuration, and so cannot be targets:\n    " ++
         T.unpack (renderPkgComponents xs) ++
@@ -637,24 +626,18 @@ configureOptsNoDir econfig bco deps isLocal package = concat
     , map ("--extra-lib-dirs=" ++) (configExtraLibDirs config)
     , maybe [] (\customGcc -> ["--with-gcc=" ++ toFilePath customGcc]) (configOverrideGccPath config)
     , ["--ghcjs" | wc == Ghcjs]
-    , ["--exact-configuration" | useExactConf]
+    , ["--exact-configuration"]
     ]
   where
     wc = view (actualCompilerVersionL.to whichCompiler) econfig
     config = view configL econfig
     bopts = bcoBuildOpts bco
 
-    -- TODO: instead always enable this when the cabal version is new
-    -- enough. That way we'll detect bugs with --exact-configuration
-    -- earlier. Cabal also might do less work then.
-    useExactConf = configAllowNewer config
-
     newerCabal = view cabalVersionL econfig >= mkVersion [1, 22]
 
     -- Unioning atop defaults is needed so that all flags are specified
     -- with --exact-configuration.
-    flags | useExactConf = packageFlags package `Map.union` packageDefaultFlags package
-          | otherwise = packageFlags package
+    flags = packageFlags package `Map.union` packageDefaultFlags package
 
     depOptions = map (uncurry toDepOption) $ Map.toList deps
       where