+++ subs/pantry/test/Pantry/HPackSpec.hs
@@ -1,19 +1,11 @@
 module Pantry.HPackSpec (spec) where
 
 import Test.Hspec
-import RIO
 import Pantry
 import Pantry.HPack
-import Path
 
 isVersion :: Version -> Bool
 isVersion _ = True
-
-customHpack :: (HasLogFunc env, HasPantryConfig env) => FilePath -> RIO env a -> RIO env a
-customHpack fp f = do
-  env <- ask
-  let a = set (pantryConfigL . hpackExecutableL) (HpackCommand fp) env
-  local (const a) f
 
 spec :: Spec
 spec = describe "Parse HPack version" $ do