+++ test/integration/tests/2997-ensure-warnings-output/files/foo/src/Foo.hs
@@ -1,1 +1,3 @@
 module Foo where
+
+foo = ()